ajustado licenças e validações

+++ src/app/api/empresas/[id]/route.ts
@@ -1,42 +1,81 @@
 import db from "@/db";
 import { empresas } from "@/db/schema";
-import { eq } from "drizzle-orm";
+import { eq, and, ne } from "drizzle-orm";
 import { NextResponse } from "next/server";
 
-export async function GET(_: Request, { params }: { params: { id: string } }) {
-  const id = Number(params.id);
+export async function GET(
+  _req: Request,
+  context: { params: { id: string } }
+) {
+  const id = Number(context.params.id);
+
+  if (isNaN(id)) {
+    return NextResponse.json({ error: "ID inválido" }, { status: 400 });
+  }
+
   const empresa = await db.select().from(empresas).where(eq(empresas.id, id));
   return NextResponse.json(empresa[0] || null);
 }
 
 export async function PUT(
   req: Request,
-  { params }: { params: { id: string } }
+  context: { params: { id: string } }
 ) {
-  const id = Number(params.id);
+  const id = Number(context.params.id);
+  if (isNaN(id)) {
+    return NextResponse.json({ error: "ID inválido" }, { status: 400 });
+  }
+
   const body = await req.json();
+
+  const {
+    razaoSocial,
+    cnpj,
+    cep,
+    cidade,
+    estado,
+    bairro,
+    complemento,
+  } = body;
+
+  const empresaComMesmoCNPJ = await db
+    .select()
+    .from(empresas)
+    .where(and(eq(empresas.cnpj, cnpj), ne(empresas.id, id)));
+
+  if (empresaComMesmoCNPJ.length > 0) {
+    return NextResponse.json(
+      { error: "Já existe uma empresa com esse CNPJ" },
+      { status: 409 }
+    );
+  }
 
   await db
     .update(empresas)
     .set({
-      razaoSocial: body.razaoSocial,
-      cnpj: body.cnpj,
-      cep: body.cep,
-      cidade: body.cidade,
-      estado: body.estado,
-      bairro: body.bairro,
-      complemento: body.complemento,
+      razaoSocial,
+      cnpj,
+      cep,
+      cidade,
+      estado,
+      bairro,
+      complemento
     })
     .where(eq(empresas.id, id));
 
-  return NextResponse.json({ message: "Empresa atualizada" });
+  return NextResponse.json({ message: "Empresa atualizada com sucesso" });
 }
 
 export async function DELETE(
-  _: Request,
-  { params }: { params: { id: string } }
+  _req: Request,
+  context: { params: { id: string } }
 ) {
-  const id = Number(params.id);
+  const id = Number(context.params.id);
+
+  if (isNaN(id)) {
+    return NextResponse.json({ error: "ID inválido" }, { status: 400 });
+  }
+
   await db.delete(empresas).where(eq(empresas.id, id));
-  return NextResponse.json({ message: "Empresa removida" });
+  return NextResponse.json({ message: "Empresa removida com sucesso" });
 }
